requirements.txt
PyNaCl>=1.2.1
